requirements.txt
# Web Framework
Flask>=2.0.0
flask-cors>=4.0.0

# Testing Framework
pytest>=7.0.0
pytest-cov>=4.0.0

# Data Processing
PyYAML>=6.0
pandas>=1.5.0
numpy>=1.24.0

# HTTP Requests (for LLM API calls)
requests>=2.28.0
aiohttp>=3.8.0

# Logging
loguru>=0.6.0

# Configuration Management
python-dotenv>=1.0.0

# Type Hints Support
typing-extensions>=4.5.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Flask
Flask>=2.0.0 